feat: scraping fans of facebook page
Fan of Facebook page is user who likes or follows this page.
If you have any of all available roles on this page you can use luster
to download list of fans as csv.

Be aware that there is limit on Facebook site which not allows to get
more then 7k users.