chore: fix invalid console input source and do not delete input lines
During testing on Windows I discovered that I used stderr instead stdin to read email
and password. This was fine on Linux but it does not work on Windows.

Additionally I decided to do not delete input lines after reading input as it was not working on Windows and it is not neaded because it does not aaffect stdout content.